fixup: merge enrichAuthHint into context.go, remove SourceUnknown, use resolved credentials path
