+++ cmd/heygen/context.go
@@ -3,14 +3,17 @@ package main
 import (
 	"context"
 	"errors"
+	"fmt"
 	"net/url"
 	"os"
+	"path/filepath"
 
 	"github.com/heygen-com/heygen-cli/internal/auth"
 	"github.com/heygen-com/heygen-cli/internal/client"
 	"github.com/heygen-com/heygen-cli/internal/config"
 	clierrors "github.com/heygen-com/heygen-cli/internal/errors"
 	"github.com/heygen-com/heygen-cli/internal/output"
+	"github.com/heygen-com/heygen-cli/internal/paths"
 	"github.com/spf13/cobra"
 )
 
@@ -53,14 +56,33 @@ func schemaFlagEnabled(cmd *cobra.Command, name string) bool {
 type credSourceKey struct{}
 
 // credSourceFromCmd retrieves the credential source stored on the command
-// context during initContext. Returns SourceUnknown if not set (e.g. skipAuth
-// commands or when credential resolution failed before storing).
+// context during initContext. Returns "" if not set (e.g. skipAuth commands
+// or when credential resolution failed before storing).
 func credSourceFromCmd(cmd *cobra.Command) auth.CredentialSource {
 	if cmd == nil {
-		return auth.SourceUnknown
+		return ""
 	}
 	src, _ := cmd.Context().Value(credSourceKey{}).(auth.CredentialSource)
 	return src
+}
+
+// enrichAuthHint sets a source-aware hint on auth errors that don't already
+// have one. Called once in the centralized error path (main.go / testutil),
+// not scattered across individual commands.
+func enrichAuthHint(cliErr *clierrors.CLIError, source auth.CredentialSource) {
+	if cliErr.ExitCode != clierrors.ExitAuth {
+		return
+	}
+	if cliErr.Hint != "" {
+		return
+	}
+	credPath := filepath.Join(paths.ConfigDir(), "credentials")
+	switch source {
+	case auth.SourceEnv:
+		cliErr.Hint = "The HEYGEN_API_KEY environment variable contains an invalid or expired key.\nGenerate a new key: https://app.heygen.com/settings/api"
+	case auth.SourceFile:
+		cliErr.Hint = fmt.Sprintf("The stored API key (%s) is invalid or expired.\nReplace it: heygen auth login\nGenerate a new key: https://app.heygen.com/settings/api", credPath)
+	}
 }
 
 // initContext sets up the config provider and, for commands that require
